refactoring code v1

+++ controllers/highlights.js
@@ -4,23 +4,30 @@ const Highlight = require('../models/highlightModel');
 module.exports.manageHighlightTrade = async (req, res) => {
    console.log(req.url)
    const { action, highlightName = 'noname' } = req.query;
-   const { orginalTradeId, highlightId } = req.params;
+   const { orginalTradeId } = req.params;
 
-   const orginalTrade = await Trade.findById(orginalTradeId).populate('highlightedTrade');
+   const doesExist = await Trade.any({ _id: orginalTradeId });
 
-   if (orginalTrade.isHighlighted == false && action == 'add') {
-      // ADDING NEW HIGHLIGHT
-      const newHighlightId = await addNewHighlight(orginalTrade, highlightName);
-      res.json({ success: true, action });
-   }
-   else if (orginalTrade.isHighlighted == true && action == 'delete') {
-      // DELETING THE HIGHLIGHT THAT IS CONNECTED TO ORGINAL TRADE ID
-      await deleteHighlight(orginalTrade);
-      res.json({ success: true, action });
-   }
-   else {
-      // IN CASE OF CONFLICT JUST FORWARDING THE RESPONSE IN ORDER TO CHANGE STATE ON THE SITE
-      res.json({ success: true, action })
+   if (doesExist) {
+
+      const orginalTrade = await Trade.findById(orginalTradeId).populate('highlightedTrade');
+
+      if (orginalTrade.isHighlighted == false && action == 'add') {
+         // ADDING NEW HIGHLIGHT
+         const newHighlightId = await addNewHighlight(orginalTrade, highlightName);
+         res.json({ success: true, action });
+      }
+      else if (orginalTrade.isHighlighted == true && action == 'delete') {
+         // DELETING THE HIGHLIGHT THAT IS CONNECTED TO ORGINAL TRADE ID
+         await deleteHighlight(orginalTrade);
+         res.json({ success: true, action });
+      }
+      else {
+         // IN CASE OF CONFLICT JUST FORWARDING THE RESPONSE IN ORDER TO CHANGE STATE ON THE SITE
+         res.json({ success: true, action })
+      }
+   } else {
+      res.json({ success: false, message: 'Orginal trade does not exist!' })
    }
 }
